Updated gerber file handling to process axis-aligned rounded rectangle regions
This is NOT a general-purpose G36/G37-delimited filled region handling implementation; it is strictly looking for rounded rectangles only, and even there only for axis-aligned ones (ie. no rotation present). A region of any other nature will generate an error and fail to import. Adding this became necessary to maintain existing levels of usability with rounded rectangle pads starting to overwhelmingly replace traditional square and obround pads in footprints.