Added new contour processing step trying to remove redundant segments after offsetting
The original GCAM contour offsetting was known to be imperfect and having difficulties as soon as offset contours changed substantially enough to require complete removal of segments shrunk to nothing. This update does NOT fully address that issue, but it does attempt to remove redundant cuts caused by the more simple cases exhibiting the problem.