Fixed a bounding box related bug that resulted in non-center rotation and mirroring points getting suggested
The rotation and the mirroring dialogs both suggest a point for those actions that is supposed to rotate or mirror a symmetrical object in-place. Calculation of that point was failing when templates were present due to an inconsistency of how they affected the calculation of the bounding boxes that the suggestions are based on.